feat(web): authjs, todos by user, validations, providers, sessions

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse, NextRequest } from 'next/server'
 import prisma from '@/lib/prisma';
 import * as yup from 'yup';
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 
 export async function GET(request: Request) { 
     const { searchParams } = new URL(request.url)
@@ -24,10 +25,14 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) { 
 
+    const user = await getUserSessionServer();
+
+    if( !user ) return NextResponse.json('Unauthorized', { status: 401 });
+
     try {
         const { complete, description } = await postSchema.validate( await request.json() );
     
-        const todo = await prisma.todo.create({ data: { complete, description } });
+        const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } });
     
         return NextResponse.json(todo); 
     } catch (error) {
@@ -36,8 +41,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) { 
+    const user = await getUserSessionServer();
+
+    if( !user ) return NextResponse.json('Unauthorized', { status: 401 });
+
     try {
-        await prisma.todo.deleteMany({ where: { complete: true }});
+        await prisma.todo.deleteMany({ where: { complete: true, userId: user.id }});
         return NextResponse.json(JSON.stringify({
             message: `Todos completed deleted`
         }), { status: 200 } );
